refactor(Main): replace Person with Wiki.Page
We can now copy a wiki page .json and paste it into the text area of our view.

When we click on the 'Parse JSON' button, we can see the changes in the Elm debugger related to the individual message and the  updated model, as well as the parsing result.