refactor: grammarString and action / actions function
Transform the matched text to uppercase before returning it.
            Ok (String.toUpper found)

The {action} in the PEG grammar doesn't directly map to Elm code. It's a placeholder that indicates where you might want to perform a custom action during parsing.

The actions function is called when a match occurs. It takes three arguments: left, found, and right. In this case, it simply returned Ok found, where found is the matched text.

The placeholder {action} in the grammar corresponds to the part of the grammar where you might want to perform a custom action. In our Elm code, this action is defined in the actions function.

If you want to perform a different action, you would modify the actions function accordingly. For example, if you wanted to transform the matched text before returning it, you would do so within the actions function.